requirements.txt
featurecloud
torch
numpy
pandas
matplotlib
scikit-learn
pyyaml